refactor: change structure of source files
