+++ src/lib/mapping/MappingItem.js
@@ -1,6 +1,6 @@
 import Logger from 'gulplog';
 import checkType from '../../util/validation';
-import NodeId from '../server/NodeId';
+import NodeId from '../ua/NodeId';
 
 
 /**
