+++ src/lib/stream/QueueStream.js
