+++ src/lib/ua/NodeId.js
