+++ test/src/lib/config/Atviseproject.spec.js
@@ -1,7 +1,7 @@
 import expect from 'unexpected';
 
 import Atviseproject from '../../../../src/lib/config/Atviseproject';
-import NodeId from '../../../../src/lib/server/NodeId';
+import NodeId from '../../../../src/lib/ua/NodeId';
 
 /** @test {Atviseproject} */
 describe('Atviseproject', function() {
