+++ test/src/transform/DisplayTransformer.spec.js
@@ -3,7 +3,7 @@ import File from 'vinyl';
 import { stub } from 'sinon';
 import expect from '../../expect';
 
-import AtviseFile from '../../../src/lib/server/AtviseFile';
+import AtviseFile from '../../../src/lib/mapping/AtviseFile';
 import { TransformDirection } from '../../../src/lib/transform/Transformer';
 import DisplayTransformer from '../../../src/transform/DisplayTransformer';
