+++ test/src/transform/Mapping.spec.js
@@ -6,8 +6,8 @@ import File from 'vinyl';
 import Logger from 'gulplog';
 import expect from '../../expect';
 import { TransformDirection } from '../../../src/lib/transform/Transformer';
-import NodeId from '../../../src/lib/server/NodeId';
-import AtviseFile from '../../../src/lib/server/AtviseFile';
+import NodeId from '../../../src/lib/ua/NodeId';
+import AtviseFile from '../../../src/lib/mapping/AtviseFile';
 
 const readFile = (path, enc, cb) => cb(null, JSON.stringify({
   typeDefinition: 'ns=1;s=VariableTypes.PROJECT.Custom',
